Refactor SKRL model configurations and agent creation
- Updated model configurations in `rover_ppo.yaml`, `rover_rpo.yaml`, `rover_sac.yaml`, `rover_td3.yaml`, and `rover_trpo.yaml` to use new model types and parameters.
- Removed the `configure_models.py` file and integrated model creation into a new `ModelFactory` class.
- Introduced a general `AgentFactory` for creating RL agents, allowing for easier agent instantiation across different algorithms.
- Registered new models and updated the model registry to support the new architecture.
- Removed deprecated code and cleaned up imports in various files.
- Updated `setup.py` to require a newer version of `isaaclab`.
- Added new test files and configuration for testing the agent creation process.
