+++ rover_envs/envs/navigation/learning/skrl/configs/rover_rpo.yaml
@@ -4,25 +4,22 @@ seed: 42
 # https://skrl.readthedocs.io/en/develop/modules/skrl.utils.model_instantiators.html
 models:
   separate: False
-  policy:  # see skrl.utils.model_instantiators.gaussian_model for parameter details
-    clip_actions: True
-    clip_log_std: True
-    min_log_std: -20.0
-    max_log_std: 2.0
-    input_shape: "Shape.STATES"
-    hiddens: [32, 32]
-    hidden_activation: ["leaky_relu", "leaky_relu", "leaky_relu"]
-    output_shape: "Shape.ACTIONS"
-    output_activation: "tanh"
-    output_scale: 1.0
-  value:  # see skrl.utils.model_instantiators.deterministic_model for parameter details
-    clip_actions: False
-    input_shape: "Shape.STATES"
-    hiddens: [32, 32]
-    hidden_activation: ["leaky_relu", "leaky_relu", "leaky_relu"]
-    output_shape: "Shape.ONE"
-    output_activation: ""
-    output_scale: 1.0
+  policy:
+    type: "GaussianNeuralNetworkConv"
+    params:
+      mlp_input_size: 5
+      mlp_layers: [256, 160, 128]
+      mlp_activation: "leaky_relu"
+      encoder_layers: [8, 16, 32, 64]
+      encoder_activation: "leaky_relu"
+  value:
+    type: "DeterministicNeuralNetworkConv"
+    params:
+      mlp_input_size: 5
+      mlp_layers: [256, 160, 128]
+      mlp_activation: "leaky_relu"
+      encoder_layers: [8, 16, 32, 64]
+      encoder_activation: "leaky_relu"
 
 
 #   learning_rate_scheduler: "KLAdaptiveRL"
