+++ rover_envs/envs/navigation/robots/aau_rover/__init__.py
@@ -2,7 +2,6 @@ import os
 
 import gymnasium as gym
 
-from ...learning.skrl import get_agent
 from . import env_cfg
 
 gym.register(
@@ -12,7 +11,6 @@ gym.register(
     kwargs={
         "env_cfg_entry_point": env_cfg.AAURoverEnvCfg,
         "best_model_path": f"{os.path.dirname(__file__)}/policies/best_agent2.pt",
-        "get_agent_fn": get_agent,
         "skrl_cfgs": {
             "PPO": f"{os.path.dirname(__file__)}/../../learning/skrl/configs/rover_ppo.yaml",
             "TRPO": f"{os.path.dirname(__file__)}/../../learning/skrl/configs/rover_trpo.yaml",
@@ -30,7 +28,6 @@ gym.register(
     kwargs={
         "env_cfg_entry_point": env_cfg.AAURoverEnvCfgSimple,
         "best_model_path": f"{os.path.dirname(__file__)}/policies/best_agent2.pt",
-        "get_agent_fn": get_agent,
         "skrl_cfgs": {
             "PPO": f"{os.path.dirname(__file__)}/../../learning/skrl/configs/rover_ppo.yaml",
             "TRPO": f"{os.path.dirname(__file__)}/../../learning/skrl/configs/rover_trpo.yaml",
@@ -48,6 +45,5 @@ gym.register(
     kwargs={
         "env_cfg_entry_point": env_cfg.AAURoverRGBResnetEnvCfg,
         #"best_model_path": f"{os.path.dirname(__file__)}/policies/best_agent.pt",
-        "get_agent_fn": get_agent,
     }
 )
