+++ setup.py
@@ -15,7 +15,7 @@ INSTALL_REQUIRES = [
     "skrl==1.4.1",
     "wandb",
     "opencv-python",
-    "isaaclab==2.0.1"
+    "isaaclab==2.1.0"
 ]
 
 # url=EXTENSION_TOML_DATA["package"]["repository"], # add later
